package: Prune actual packages from packages vector
When querying packages and filtering using the --filter option, we
currently only filter when we print the results. Later, however, we end
up printing statistics for the full vector of packages. This is
confusing behavior -- a user that's intentionally filtering by specific
packages probably only cares about those packages. Let's instead add
logic to prune the vector of packages so that the packages passed to
print_results() are the ones we actually want to count and print.

Signed-off-by: David Vernet <void@manifault.com>